test(react-query): move type only tests
