+++ packages/react-query/src/__tests__/useQueries.test.tsx
@@ -1621,65 +1621,4 @@ describe('useQueries', () => {
       ),
     )
   })
-
-  it('should return correct data for dynamic queries with mixed result types', async () => {
-    const key1 = queryKey()
-    const key2 = queryKey()
-
-    function Page() {
-      const Queries1 = {
-        get: () =>
-          queryOptions({
-            queryKey: key1,
-            queryFn: async () => {
-              await sleep(10)
-              return 1
-            },
-          }),
-      }
-      const Queries2 = {
-        get: () =>
-          queryOptions({
-            queryKey: key2,
-            queryFn: async () => {
-              await sleep(10)
-              return true
-            },
-          }),
-      }
-
-      const queries1List = [1, 2, 3].map(() => ({ ...Queries1.get() }))
-      const result = useQueries({
-        queries: [...queries1List, { ...Queries2.get() }],
-      })
-
-      expectTypeOf(result).toEqualTypeOf<
-        [
-          ...Array<UseQueryResult<number, Error>>,
-          UseQueryResult<boolean, Error>,
-        ]
-      >()
-
-      expectTypeOf(result[0]?.data).toEqualTypeOf<
-        number | boolean | undefined
-      >()
-
-      return (
-        <div>
-          <div>
-            data1: {String(result[0]?.data ?? 'null')}, data2:{' '}
-            {String(result[1]?.data ?? 'null')}, data3:{' '}
-            {String(result[2]?.data ?? 'null')}, data4:{' '}
-            {String(result[3]?.data ?? 'null')}
-          </div>
-        </div>
-      )
-    }
-
-    const rendered = renderWithClient(queryClient, <Page />)
-
-    await waitFor(() =>
-      rendered.getByText('data1: 1, data2: 1, data3: 1, data4: true'),
-    )
-  })
 })
