+++ packages/react-query/src/__tests__/useSuspenseQueries.test-d.tsx
@@ -1,4 +1,5 @@
 import { describe, expectTypeOf, it } from 'vitest'
+import { queryKey, sleep } from 'src/__tests__/utils'
 import { skipToken, useSuspenseQueries } from '..'
 import { queryOptions } from '../queryOptions'
 import type { OmitKeyof } from '..'
@@ -153,5 +154,45 @@ describe('UseSuspenseQueries config object overload', () => {
 
       expectTypeOf(data).toEqualTypeOf<Data>()
     })
+  })
+
+  it('should return correct data for dynamic queries with mixed result types', () => {
+    const key1 = queryKey()
+    const key2 = queryKey()
+
+    const Queries1 = {
+      get: () =>
+        queryOptions({
+          queryKey: key1,
+          queryFn: async () => {
+            await sleep(10)
+            return 1
+          },
+        }),
+    }
+    const Queries2 = {
+      get: () =>
+        queryOptions({
+          queryKey: key2,
+          queryFn: async () => {
+            await sleep(10)
+            return true
+          },
+        }),
+    }
+
+    const queries1List = [1, 2, 3].map(() => ({ ...Queries1.get() }))
+    const result = useSuspenseQueries({
+      queries: [...queries1List, { ...Queries2.get() }],
+    })
+
+    expectTypeOf(result).toEqualTypeOf<
+      [
+        ...Array<UseSuspenseQueryResult<number, Error>>,
+        UseSuspenseQueryResult<boolean, Error>,
+      ]
+    >()
+
+    expectTypeOf(result[0]?.data).toEqualTypeOf<number | boolean>()
   })
 })
