+++ packages/react-query/src/__tests__/useSuspenseQueries.test.tsx
@@ -4,21 +4,15 @@ import {
   beforeAll,
   describe,
   expect,
-  expectTypeOf,
   it,
   vi,
 } from 'vitest'
 import { act, fireEvent, render, waitFor } from '@testing-library/react'
 import * as React from 'react'
 import { ErrorBoundary } from 'react-error-boundary'
-import {
-  queryOptions,
-  skipToken,
-  useSuspenseQueries,
-  useSuspenseQuery,
-} from '..'
+import { skipToken, useSuspenseQueries, useSuspenseQuery } from '..'
 import { createQueryClient, queryKey, renderWithClient, sleep } from './utils'
-import type { UseSuspenseQueryOptions, UseSuspenseQueryResult } from '..'
+import type { UseSuspenseQueryOptions } from '..'
 
 type NumberQueryOptions = UseSuspenseQueryOptions<number>
 
@@ -698,64 +692,5 @@ describe('useSuspenseQueries 2', () => {
       'skipToken is not allowed for useSuspenseQueries',
     )
     consoleErrorSpy.mockRestore()
-  })
-
-  it('should return correct data for dynamic queries with mixed result types', async () => {
-    const key1 = queryKey()
-    const key2 = queryKey()
-
-    function Page() {
-      const Queries1 = {
-        get: () =>
-          queryOptions({
-            queryKey: key1,
-            queryFn: async () => {
-              await sleep(10)
-              return 1
-            },
-          }),
-      }
-      const Queries2 = {
-        get: () =>
-          queryOptions({
-            queryKey: key2,
-            queryFn: async () => {
-              await sleep(10)
-              return true
-            },
-          }),
-      }
-
-      const queries1List = [1, 2, 3].map(() => ({ ...Queries1.get() }))
-      const result = useSuspenseQueries({
-        queries: [...queries1List, { ...Queries2.get() }],
-      })
-
-      expectTypeOf(result).toEqualTypeOf<
-        [
-          ...Array<UseSuspenseQueryResult<number, Error>>,
-          UseSuspenseQueryResult<boolean, Error>,
-        ]
-      >()
-
-      expectTypeOf(result[0]?.data).toEqualTypeOf<number | boolean>()
-
-      return (
-        <React.Suspense fallback="loading...">
-          <div>
-            data1: {String(result[0]?.data ?? 'null')}, data2:{' '}
-            {String(result[1]?.data ?? 'null')}, data3:{' '}
-            {String(result[2]?.data ?? 'null')}, data4:{' '}
-            {String(result[3]?.data ?? 'null')}
-          </div>
-        </React.Suspense>
-      )
-    }
-
-    const rendered = renderWithClient(queryClient, <Page />)
-
-    await waitFor(() =>
-      rendered.getByText('data1: 1, data2: 1, data3: 1, data4: true'),
-    )
   })
 })
